requirements.txt
Flask>=3.0.0,<4.0.0
pydantic>=2.5.0,<3.0.0
PyYAML>=6.0
requests>=2.31.0
pytest>=7.4.0
